reorganize functions into more relevant componenets
